fail build if diff exists
Right now the build tries to detect mismatches between what is in the
change versus any changes actually occurred from a clean rebuild (e.g.
missed checking in any updates to generated code). However if those are
present, the build logs print that out but do not cause the build to
fail.

Also bumps GOVERSION to match what is used in Dockerfile so we don't
have go.mod or go.sum as a reason for above.

Signed-off-by: Bob Callaway <bob.callaway@gmail.com>