types/gitpush: Implement git push certificates
This is an initial attempt at implementing support for git push
certificates in rekor.

Git push certificates allows the committer to sign the commits they
pushed towards some remote. These are important to log as they can work
as a canary if there are malicious commits on the git forge server. An
example of this previously is the git.php.net compromise along with
several attempts at compromising kernel.org in the past.

kernel.org already implements a transparency logs with all git push
operations, but not all of them are signed[0].

There is also a POC implementation of a monitor for this transparency
log[1], but it would be better served to be on shared infrastructure
with other projects instead of being reinvented on a per project basis.

The current implementation is a veryvery naive json schema over the
documented push certificate mostly representing everything as strings
with minimal validation over the schema[2]. Hopefully this can be
better validated with parsing support in something like go-git in the
future[3]

It currently missing some frontend tool for the parsing itself, but the
idea is to hook it up with a post-receive hook from the repository which
would feed rekor[4].

[0]: https://korg.docs.kernel.org/gitolite/transparency-log.html
[1]: https://tlog.linderud.dev/
[2]: https://git-scm.com/docs/pack-protocol/2.29.0#_push_certificate
[3]: https://github.com/go-git/go-git
[4]: https://git-scm.com/docs/git-receive-pack/2.11.4#_post_receive_hook

Signed-off-by: Morten Linderud <morten@linderud.pw>