Ensure envelope hash is computed at validation time
Previously the digest of the envelope contents was computed before
returning the canonicalized byte stream that is inserted into the log,
but was not stored in the internal object representation so it was not
available when IndexKeys was called. This fixes it to be available, and
also has the benefit of ignoring the client provided value (as we should
be relying only on server side computations for these keys).

Signed-off-by: Bob Callaway <bcallaway@google.com>