Feature: add search support for sha512 (#1142)
* Feature: add search support for sha512

This commit introduces the possibility to *search* via sha512 digests.
Example of how this could be useful is for the npm ecosystem. npm often relies
on sha512 digests, and provenance attestations in intoto format may use
sha512 digests as the subject. Subjects are extract during uploads and added
as index keys, which are then stored in Redis. This feature lets the cli
accept sha512 digests, and the server to accept them to use when searching in
the Redis cache.

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>

* Added SHA512 to the cli help

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>

* Re-ran code generation.

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>

* Added e2e tests that uploads an intoto attestation, with a SHA512 subject.
The tests then verifies that the entry can be found via the SHA512 digest.

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>