Fix panic when parsing SSH SK pubkeys
This is because these pubkey types do not implement
ssh.CryptoPublicKey, which causes a panic when we try to do
a type assertion.

Also realized we weren't handling SSH certs, so now we extract the
pubkey from the cert before trying to parse it.

Had to reimplement parsing the SK pubkeys because there is no other way
to extract the raw pubkey from it. After https://github.com/golang/go/issues/62518,
this will get cleaned up.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>