Add index performance script (#2042)
Add terraform configuration and scripts to set up rekor standalone on
GCP, perform a series of insert and search operations, use Prometheus to
gather metrics, and plot the results with gnuplot.

The scripts added here are for comparing mysql and redis as index
storage backends. Other types of performance measurement scripts could
be added here in the future.

To get a realistic sense of query speed for searches, a large data set
is needed. Rather than using the rekor API to insert real data, fake
data is generated and uploaded directly to the backend before searching
it.

Different types of searches are performed: searches where there should
be many results, searches where there should be few results, and
searches where there should be no results. The goal is not to compare
the latency of these different searches, but to take the overall average
to compare across backends.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>