Add script to delete indexes from Redis
Add a script to check what indexes are in the EntryIndex table in a
MySQL database and deletes those indexes from the Redis database. This
is meant to be done after Rekor is migrated to using
--search_index.storage_provider=mysql and the MySQL table has been
backfilled with all the indexes. Deleting the data allows us to downsize
the Redis instance.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>