Increase rekor client retries in backfill script
Increase the RetryCount setting from the default of 3 up to 10 in order
to avoid giving up too quickly when the script hits the server rate
limit. The retryablehttp client does an exponential backoff, so
increasing the number of tries also increases the amount of time it will
wait in between each try before it eventually succeeds.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>