Unset DisableKeepalive for backfill HTTP client (#2137)
* Unset DisableKeepalive for backfill HTTP client

Disabling Keep-Alive, as done by the default transport setting in the
hashicorp cleanhttp package, seems to conflict with a network setting
between the public good Rekor instances and the bastion and results in
GET requests stalling or timing out after processing a few entries. This
change adds an option to the rekor client to unset the DisableKeepalive
setting and has the backfill script utilize that option. Other uses of
the rekor client will see no change.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Increase rekor client retries in backfill script

Increase the RetryCount setting from the default of 3 up to 10 in order
to avoid giving up too quickly when the script hits the server rate
limit. The retryablehttp client does an exponential backoff, so
increasing the number of tries also increases the amount of time it will
wait in between each try before it eventually succeeds.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

---------

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>