Fix index cleanup script for mixed case keys
Some index keys, especially workload identifiers, have uppercase
characters. MySQL is case insensitive, but Redis is not. The index
storage backends account for this by converting lookup keys to lowercase
for Redis queries, but the cleanup script did not account for this and
was leaving some entries present in MySQL undeleted in Redis because
they appeared to be different strings. This change ensures that these
keys will be deleted from Redis when they are fully migrated to MySQL,
even if the case is different.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>