chore: relax go directive to permit 1.22.x
Setting this to 1.23.4 requires all consumers to be building with the
absolute latest 1.23.4 or newer release of Go 1.23 and to update their
own go.mod accordingly — this seems unnecessarily restrictive for a
library module, particularly as the code itself doesn't currently use
any modern language constructs and builds fine even with older Go
versions.

Instead set the go directive to 1.22.0 and use the toolchain directive
to recommend the latest 1.23.x when building locally.

Note: this also upgrades google/trillian to v1.7.1 and downgrades
sigs.k8s.io/release-utils to v0.8.4, which are both required in order to
allow the go directive to be relaxed here.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>