Support per-shard signing keys (#2330)
* Support per-shard signing keys

This change enables key rotation with a per-shard
signing key configuration. The LogRanges structure
now holds both active and inactive shards, with the
LogRange structure containing a signer, encoded
public key and log ID based on the public key.

This change is backwards compatible. If no signing
configuration is specified, the active shard
signing configuration is used for all shards.

Minor change: Standardized log ID vs tree ID, where
the former is the pubkey hash and the latter is the
ID for the Trillian tree.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* resolve codeql, remove key password from string

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Update rekor-cli to pass tree ID for verification

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Fix range printing

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* suppress codeql

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* remove lgtm, it does nothing

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* address comments

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Apply suggestions from code review

Co-authored-by: Bob Callaway <bobcallaway@users.noreply.github.com>
Signed-off-by: Hayden B <hblauzvern@google.com>

---------

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>
Signed-off-by: Hayden B <hblauzvern@google.com>
Co-authored-by: Bob Callaway <bobcallaway@users.noreply.github.com>