Disable stable checkpoint by default
The current default setting means that deployers are required to spin up
a Redis instance alongside Rekor. With this now disabled by default,
deployers can choose whether they actually want to use stable
checkpoints.

I'm also tempted to remove the stable checkpoint feature entirely
because I'm not sure anyone uses it in the intended way to enable
witnessing, but since it's not supported in Rekor v2, I'll leave this
as-is to avoid a breaking change.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>