pkg: separate pki types from implementations
pkg/pki package defines both the interface types for PublicKey
and Signature, linked to many external packages, and also all the
implementations for pki via static factory map.

This separates the types to separate package so the packages
that use them can be included without a big dependency chain.
The types are aliased to the old pkg/pki package so that this
change wouldn't break any backwards compatibility.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>